tests/wpt/web-platform-tests/offscreen-canvas/the-offscreen-canvas/size.attributes.idl.set.zero.html

<!DOCTYPE html>
<!-- DO NOT EDIT! This test has been generated by tools/gentest.py. -->
<title>OffscreenCanvas test: size.attributes.idl.set.zero</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/common/canvas-tests.js"></script>

<h1>size.attributes.idl.set.zero</h1>
<p class="desc">Setting width/height IDL attributes to 0</p>


<script>
var t = async_test("Setting width/height IDL attributes to 0");
t.step(function() {

var offscreenCanvas = new OffscreenCanvas(100, 50);
var ctx = offscreenCanvas.getContext('2d');

offscreenCanvas.width = 0;
offscreenCanvas.height = 0;
_assertSame(offscreenCanvas.width, 0, "offscreenCanvas.width", "0");
_assertSame(offscreenCanvas.height, 0, "offscreenCanvas.height", "0");

t.done();

});
</script>
